Move `Number` to own module and cleanup code
